+++ src/components/Home/Home.vue
@@ -126,9 +126,7 @@ export default {
     }
 
     .bmc-button:hover, .bmc-button:active, .bmc-button:focus {
-        -webkit-box-shadow: 0px 1px 2px 2px rgba(190, 190, 190, 0.5) !important;
         text-decoration: none !important;
-        box-shadow: 0px 1px 2px 2px rgba(190, 190, 190, 0.5) !important;
         opacity: 0.85 !important;
         color: #FFFFFF !important;
     }